Refactor: Move assertCanonicalDepositTxShape and related methods from TradeDataValidation to DepositTxValidation
